api-react/requirements.txt
numpy==1.23.5
flask==2.2.5
scikit-learn==1.2.2
pandas==1.5.3
scipy==1.10.1
requests==2.28.2
streamlit==1.20.0
werkzeug==2.2.3
flask_cors==3.0.10
python-dotenv==1.0.1